GROOV-37 tdl Main 최종 UI작업
