+++ package.json
@@ -32,6 +32,7 @@
     "install": "^0.13.0",
     "moment": "^2.29.4",
     "net": "^1.0.2",
+    "path-browserify": "^1.0.1",
     "quill": "^1.3.6",
     "react": "^18.2.0",
     "react-bootstrap": "^2.9.1",
