+++ src/App.css
@@ -13,14 +13,17 @@ body {
 
 .SlideContainer {
     width: 250px;
-    height: 100vh;
     background-color: rgb(239, 244, 252);
     padding-left: 20px;
     padding-right: 20px;
     z-index: -1;
     margin: 0px;
     float: left;
-    overflow: auto;
+    overflow-y: scroll;
+}
+
+.SlideContainer::-webkit-scrollbar {
+    display: none;
 }
 
 .MainContainer {
