+++ src/pages/Main/components/SlideBar/SlideBar/ToDoList/ToDoListSlide.module.css
@@ -57,11 +57,52 @@ a {
     font-weight: bold;
 }
 
-.colorbox {
+.tdlul {
+    width: 100%;
+    list-style: none;
+    padding: 0px;
+    padding-bottom: 60px;
+}
+
+.tdlli {
+    margin-bottom: 5px;
+}
+
+.tdlDBInsert {
+    height: 40px;
+    position: relative;
+    cursor: pointer;
+}
+
+.tdltitle {
+    width: 90%;
+    height: 18px;
+    color: white;
+    font-size: 13px;
+    font-weight: 700;
+    position: relative;
+    top: 2px;
+    left: 2px;
+    overflow: hidden;
+}
+
+.starimg {
     width: 20px;
     height: 20px;
-    background-color: black;
-    float: left;
-    margin-top: 2px;
-    margin-right: 5px;
+    position: absolute;
+    top: 5px;
+    right: -20px;
+}
+
+.tdlDBInsert:hover .starimg {
+    right: -2px;
+}
+
+.starimg:active {
+    transform: scale(0.8);
+}
+
+.starActive {
+    top: 5px;
+    right: -2px;
 }
