+++ src/pages/ToDoList/ToDoListMain/ToDoListMain.module.css
@@ -87,6 +87,7 @@
 }
 
 .tdltitle {
+    width: 90%;
     color: white;
     font-size: 15px;
     font-weight: 700;
